+++ src/BayesFilters/include/BayesFilters/Gaussian.h
@@ -22,11 +22,15 @@ namespace bfl {
 class bfl::Gaussian : public bfl::GaussianMixture
 {
 public:
-    Gaussian();
+    Gaussian() noexcept;
 
-    Gaussian(const std::size_t dim_linear);
+    Gaussian(const std::size_t dim_linear) noexcept;
 
-    Gaussian(const std::size_t dim_linear, const std::size_t dim_circular);
+    Gaussian(const std::size_t dim_linear, const std::size_t dim_circular) noexcept;
+
+    virtual ~Gaussian() noexcept;
+
+    void resize(const std::size_t dim_linear, const std::size_t dim_circular = 0);
 
     /**
      * Non-virtual methods of GaussianMixture are overriden here
@@ -54,8 +58,6 @@ public:
     double& weight();
 
     const double& weight() const;
-
-    virtual ~Gaussian() noexcept { };
 };
 
 #endif /* GAUSSIAN_H */