+++ src/BayesFilters/include/BayesFilters/ParticleSet.h
@@ -20,11 +20,15 @@ namespace bfl {
 class bfl::ParticleSet : public bfl::GaussianMixture
 {
 public:
+    ParticleSet() noexcept;
+
     ParticleSet(const std::size_t components, const std::size_t dim) noexcept;
 
     ParticleSet(const std::size_t components, const std::size_t dim_linear, const std::size_t dim_circular) noexcept;
 
     virtual ~ParticleSet() noexcept;
+
+    void resize(const std::size_t components, const std::size_t dim_linear, const std::size_t dim_circular = 0) override;
 
     ParticleSet& operator+=(const ParticleSet& rhs);